+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+Release v1.5.2 (2016-11-03)
+===
+
+Service Client Updates
+---
+`service/directconnect`: Updates service API and documentation
+`service/endpoints.json`: Adds new service
+
 Release v1.5.1 (2016-11-02)
 ===